Add lomaji inputplugin
The script print_lomaji.py generates the table in the inputplugin
file. It is a very naive script that simply concatenates starting
consonants with ending tonal phonemes based on a simple set of
rules. The generated sequences should be a superset of all the
tokens we would want to generate, but it also allows the user to
type large number of tokens with no meaning in the language.