Add hanji inputplugin
This plugin builds on the lomaji plugin by using the same input
sequences for single-token words. Multi-token words are simplified
so that the user only has to enter the atonal tokens separated by
hyphens to generated the desired compound hanji. The list of hanji
vocabulary is taken from the main word entry data file from the
moedict twblg project.